Changes to be committed: 	deleted:    ResourceCentre_23046904/inventory/__pycache__/__init__.cpython-38.pyc 	deleted:    ResourceCentre_23046904/inventory/__pycache__/camera.cpython-38.pyc 	deleted:    ResourceCentre_23046904/inventory/__pycache__/inventory.cpython-38.pyc 	deleted:    ResourceCentre_23046904/inventory/__pycache__/laptop.cpython-38.pyc 	deleted:    ResourceCentre_23046904/teamproject/readme.txt 	deleted:    ResourceCentre_23046904/tests/__pycache__/__init__.cpython-38.pyc 	deleted:    ResourceCentre_23046904/tests/__pycache__/test_US_01.cpython-38-pytest-8.2.1.pyc 	renamed:    ResourceCentre_23046904/inventory/__init__.py -> inventory/__init__.py 	modified:   inventory/__pycache__/__init__.cpython-38.pyc 	modified:   inventory/__pycache__/camera.cpython-38.pyc 	modified:   inventory/__pycache__/inventory.cpython-38.pyc 	modified:   inventory/__pycache__/laptop.cpython-38.pyc 	renamed:    ResourceCentre_23046904/inventory/camera.py -> inventory/camera.py 	renamed:    ResourceCentre_23046904/inventory/inventory.py -> inventory/inventory.py 	renamed:    ResourceCentre_23046904/inventory/laptop.py -> inventory/laptop.py 	deleted:    readme.txt 	renamed:    ResourceCentre_23046904/resourcecentre.py -> resourcecentre.py 	renamed:    ResourceCentre_23046904/tests/__init__.py -> tests/__init__.py 	modified:   tests/__pycache__/__init__.cpython-38.pyc 	modified:   tests/__pycache__/test_US_01.cpython-38-pytest-8.2.1.pyc 	renamed:    ResourceCentre_23046904/tests/test_US_01.py -> tests/test_US_01.py
