--- readme.txt
@@ -1,4 +1,0 @@
-initial commit
-
-
-jeremy created feature 2
